+++ README.md
@@ -28,6 +28,12 @@ Pims relies on the Bioformats [loci_tools.jar](http://downloads.openmicroscopy.o
 and downloads this at program runtime.
 * The output directory location must not include any spaces in the path. Eg: `/Documents/path/to/output/` is fine, but `Documents/folder with spaces/to/output/` is not.
 
+![podo screen shot 2018-05-23 at 4 50 11 pm](https://user-images.githubusercontent.com/30920819/48197692-98110d80-e3aa-11e8-9f85-aba0b1d5dd49.jpg)
+
+
+![podo screen shot 2018-05-23 at 5 31 56 pm](https://user-images.githubusercontent.com/30920819/48197735-becf4400-e3aa-11e8-8a07-7dcc852799e5.jpg)
+
+
 ## For developers
 ### Setup development environment
 
@@ -45,10 +51,10 @@ pip install -e .
 pythonw podocytes/main.py
 ```
 
-
 ### Running PyInstaller to create macOS build
 
-```python
+```
 pip install -e .
 pyinstaller Podo.spec -w -F -y
 ```
+
