+++ podocytes/__init__.py
@@ -1,0 +1,15 @@
+"""Podocyte automated analysis program
+
+Load, segment, count, and measure glomeruli and podocytes in
+fluorescence images.
+
+The main package of ``podocytes`` provides the base functionality for
+counting podocytes within the kidney glomeruli.
+
+Subpackages
+-----------
+validate
+
+"""
+
+from podocytes._version import __version__
