+++ podocytes/util.py
@@ -1,3 +1,13 @@
+import os
+import time
+import logging
+
+from gooey.python_bindings.gooey_decorator import Gooey as gooey
+from gooey.python_bindings.gooey_parser import GooeyParser
+
+from podocytes.__init__ import __version__
+
+
 def find_files(input_directory, ext):
     """Recursive search for filenames matching specified extension.
 
@@ -21,3 +31,60 @@ def find_files(input_directory, ext):
                 filename = os.path.join(root, f)
                 filelist.append(filename)
     return filelist
+
+
+def configure_parser_default(parser):
+    parser.add_argument('input_directory', widget='DirChooser',
+                        help='Folder containing files for processing.')
+    parser.add_argument('output_directory', widget='DirChooser',
+                        help='Folder to save output analysis files.')
+    parser.add_argument('glomeruli_channel_number',
+                        help='Fluorescence channel with glomeruli.',
+                        type=int, default=1)
+    parser.add_argument('podocyte_channel_number',
+                        help='Fluorescence channel with podocytes.',
+                        type=int, default=2)
+    parser.add_argument('minimum_glomerular_diameter',
+                        help='Minimum glomerular diameter (microns).',
+                        type=float, default=30)
+    parser.add_argument('maximum_glomerular_diameter',
+                        help='Maximum glomerular diameter (microns).',
+                        type=float, default=300)
+    parser.add_argument('file_extension',
+                        help='Extension of image file format (.tif, etc.)',
+                        type=str, default='.lif')
+    return parser
+
+
+def log_file_begins(args, timestamp):
+    """Initialize logging and begin writing log file.
+
+    Parameters
+    ----------
+    args : user input arguments
+        Input arguments from user.
+    timestamp :  str
+        Local time as string formatted as day-month-year_hour-minute-AM/PM
+
+    Returns
+    -------
+    log_filename : str
+        Filepath to output log text file location.
+    """
+    log_filename = os.path.join(args.output_directory, f"log_podo_{timestamp}")
+    logging.basicConfig(
+        format="%(asctime)s %(message)s",
+        level=logging.DEBUG,
+        handlers=[
+            logging.FileHandler(f"{log_filename}.log"),
+            logging.StreamHandler()
+        ])
+    # Log user input arguments
+    logging.info(f"Podocyte automated analysis program, version {__version__}")
+    logging.info(f"{timestamp}")
+    logging.info("========== USER INPUT ARGUMENTS ==========")
+    user_input = vars(args)
+    for key, val in user_input.items():
+        logging.info(f"{key}: {val}")
+    logging.info("======= END OF USER INPUT ARGUMENTS =======")
+    return log_filename
